Refactor and add documentation to AsyncHelper methods
Generalize the parameter, implementation, and two invocations of GetModifiersWithAsyncKeywordAdded to accept any SyntaxTokenList of modifiers, in order to facilitate performing this operation for anonymous function modifiers and local function modifiers as well as method modifiers.

Add handling for non-void/awaitable return types T in GetAsyncReturnType in order to correctly convert to Task<T> instead of Task.